Do not let NothingToDoJob pop out to the job worker
This exception cancels and set the whole job to done. This is a risky
and unexpected behavior when we have cascading imports. Such as an
import of product which triggers a import of product category, if the
import of the product category raises a NothingToDoJob, the whole job
including the import of product would be canceled.